pyproject.toml
[tool.poetry]
name = "baal"
version = "1.9.2"
description = "Library to enable Bayesian active learning in your research or labeling work."
authors = ["Parmida Atighehchian <parmida.atighehchian@servicenow.com>",
           "Frederic Branchaud-Charron <frederic.branchaud-charron@servicenow.com>",
           "Jan Freyberg <jan.freyberg@gmail.com>",
           "Lorne Schell <lorne.schell@servicenow.com>",
           "Rafael Pardina <rafael.pardinas@servicenow.com>"
           ]
license = "Apache-2.0"
readme = "README.md"
documentation = "https://baal.readthedocs.io"
repository = "https://github.com/ElementAI/baal/"

[tool.poetry.dependencies]
python = ">=3.8,<4"
torch = ">=1.6.0"
torchmetrics = "^0.9.3"
h5py = "^3.4.0"
numpy = "^1.21.2"
matplotlib = "^3.4.3"
Pillow = ">=6.2.0"
tqdm = "^4.62.2"
structlog = "^22.0.0"
scikit-learn = "^1.0.0"
scipy = "^1.7.1"

# Vision
torchvision = { version = ">=0.7.0", optional=true }
lightning-flash = { version = ">=0.7.5", optional=true }

# NLP
transformers = {version = ">=4.10.2", optional=true}
accelerate = {version = "^0.28.0", optional=true}
datasets = {version = ">=2.14.6", optional=true}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
torch-hypothesis = "0.2.0"
hypothesis = "4.24.0"
flake8 = "^3.9.2"
pytest-mock = "^3.6.1"
black = "^22.3.0"
# Issue with mypy https://github.com/pydantic/pydantic/issues/5192
mypy = "<=1.0.1"
bandit = "^1.7.1"

# Documentation
docutils = "0.16"

# Documentation
mkdocs-jupyter = "^0.21.0"
mkdocs-material = "^8.5.6"
Pygments = "^2.12.0"
mkdocstrings = {extras = ["python"], version = "^0.18.1"}
mkdocs-exclude-search = "^0.6.4"

[tool.poetry.extras]
vision = ["torchvision", "lightning-flash"]
nlp = ["transformers", "datasets", "accelerate"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
